WEB_SCRAPPER/requirements.txt
beautifulsoup4==4.12.3
gspread==6.1.2
google-auth==2.35.0
pandas==2.2.3
requests==2.32.3
